requirements.txt
git+https://github.com/rayer4u/aldryn-newsblog.git@test2
cmsplugin-filer
coreapi
Django<2
django-classy-tags
django-cms
django-easy-thumbnails-watermark
django-filer
django-filter
django-registration
django-sekizai
django-taggit-serializer
django-treebeard
djangocms-admin-style
djangocms-link
djangocms-style
# djangocms-googlemap
djangocms-snippet
djangocms-text-ckeditor
djangocms-video
#djangocms-column
djangorestframework
drf-yasg
easy_thumbnails
flex
gunicorn
html5lib==0.999999  # djangocms-text-ckeditor 3.6.0 still needed
Pillow
six
